// src/pokeapi/generated.ts
export interface components {
  schemas: {
// ...
    EvolutionChainDetail: {
      readonly id: number;
      baby_trigger_item: components["schemas"]["ItemSummary"];
      readonly chain: {
        evolution_details: unknown[];
        evolves_to: {
          evolution_details: {
            version_group: {
              /** @example 1 */
              name: string;
              /**
               * Format: uri
               * @example 2
               */
              url: string;
            };
            is_default: boolean;
            gender: {
              /** @example 1 */
              name: string;
              /**
               * Format: uri
               * @example 2
               */
              url: string;
            } | null;
            held_item: {
              /** @example 1 */
              name: string;
              /**
               * Format: uri
               * @example 2
               */
              url: string;
            } | null;
            item: {
              /** @example 1 */
              name: string;
              /**
               * Format: uri
               * @example 2
               */
              url: string;
            } | null;
            known_move: Record<string, never> | null;
            known_move_type: Record<string, never> | null;
            location: {
              name: string;
              /** Format: uri */
              url: string;
            } | null;
            /** Format: int32 */
            min_affection: number | null;
            /** Format: int32 */
            min_beauty: number | null;
            /** Format: int32 */
            min_damage_taken: number | null;
            /** Format: int32 */
            min_happiness: number | null;
            /** Format: int32 */
            min_level: number | null;
            /** Format: int32 */
            min_move_count: number | null;
            /** Format: int32 */
            min_steps: number | null;
            near_special_rock: boolean | null;
            needs_multiplayer: boolean | null;
            needs_overworld_rain: boolean | null;
            party_species: string | null;
            party_type: string | null;
            relative_physical_stats: string | null;
            time_of_day: string;
            trade_species: string | null;
            trigger: {
              name: string;
              /** Format: uri */
              url: string;
            };
            turn_upside_down: boolean;
            used_move: Record<string, never> | null;
            region: {
              name: string;
              /** Format: uri */
              url: string;
            } | null;
            base_form: {
              name: string;
              /** Format: uri */
              url: string;
            } | null;
            evolved_form: {
              name: string;
              /** Format: uri */
              url: string;
            } | null;
          }[];
          is_baby: boolean;
          species: {
            /** @example happiny */
            name: string;
            /**
             * Format: uri
             * @example https://pokeapi.co/api/v2/pokemon-species/440/
             */
            url: string;
          };
        }[];
        is_baby: boolean;
        species: {
          /** @example happiny */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/pokemon-species/440/
           */
          url: string;
        };
      };
    };
// ...
    PokemonDetail: {
// ...
      readonly past_abilities: {
        abilities: {
          ability: {
            /** @example levitate */
            name: string;
            /**
             * Format: uri
             * @example https://pokeapi.co/api/v2/ability/26/
             */
            url: string;
          };
          is_hidden: boolean;
          /**
           * Format: int32
           * @example 1
           */
          slot: number;
        }[];
        generation: {
          /** @example generation-vi */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/generation/6/
           */
          url: string;
        };
      }[];
// ...
      readonly past_stats: {
        generation: {
          /** @example generation-vi */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/generation/6/
           */
          url: string;
        };
        stats: {
          /**
           * Format: int32
           * @example 45
           */
          base_stat: number;
          /**
           * Format: int32
           * @example 0
           */
          effort: number;
          stat: {
            /** @example speed */
            name: string;
            /**
             * Format: uri
             * @example https://pokeapi.co/api/v2/stat/6/
             */
            url: string;
          };
        }[];
      }[];
      readonly types: {
        /**
         * Format: int32
         * @example 1
         */
        slot: number;
        type: {
          /** @example ghost */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/type/8/
           */
          url: string;
        };
      }[];
      readonly past_types: {
        generation: {
          /** @example generation-v */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/generation/5/
           */
          url: string;
        };
        types: {
          /**
           * Format: int32
           * @example 1
           */
          slot: number;
          type: {
            /** @example normal */
            name: string;
            /**
             * Format: uri
             * @example https://pokeapi.co/api/v2/type/1/
             */
            url: string;
          };
        }[];
      }[];
    };
// ...
    PokemonSpeciesDetail: {
      readonly id: number;
      name: string;
      order?: number | null;
      gender_rate?: number | null;
      capture_rate?: number | null;
      base_happiness?: number | null;
      is_baby?: boolean;
      is_legendary?: boolean;
      is_mythical?: boolean;
      hatch_counter?: number | null;
      has_gender_differences?: boolean;
      forms_switchable?: boolean;
      growth_rate: components["schemas"]["GrowthRateSummary"];
      readonly pokedex_numbers: components["schemas"]["PokemonDexEntry"][];
      readonly egg_groups: {
        /** @example monster */
        name: string;
        /**
         * Format: uri
         * @example https://pokeapi.co/api/v2/egg-group/1/
         */
        url: string;
      }[];
      color: components["schemas"]["PokemonColorSummary"];
      shape: components["schemas"]["PokemonShapeSummary"];
      evolves_from_species: components["schemas"]["PokemonSpeciesSummary"];
      evolution_chain: components["schemas"]["EvolutionChainSummary"];
      habitat: components["schemas"]["PokemonHabitatSummary"];
      generation: components["schemas"]["GenerationSummary"];
      readonly names: {
        language: {
          /** @example en */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/language/9/
           */
          url: string;
        };
        /** @example bulbasaur */
        name: string;
      }[];
      readonly pal_park_encounters: {
        area: {
          /** @example field */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/pal-park-area/2/
           */
          url: string;
        };
        /**
         * Format: int32
         * @example 50
         */
        base_score: number;
        /**
         * Format: int32
         * @example 30
         */
        rate: number;
      }[];
      readonly form_descriptions: components["schemas"]["PokemonSpeciesDescription"][];
      readonly flavor_text_entries: components["schemas"]["PokemonSpeciesFlavorText"][];
      readonly genera: {
        /** @example Seed Pokémon */
        genus: string;
        language: {
          /** @example en */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/language/9/
           */
          url: string;
        };
      }[];
      readonly varieties: {
        is_default: boolean;
        pokemon: {
          /** @example bulbasaur */
          name: string;
          /**
           * Format: uri
           * @example https://pokeapi.co/api/v2/pokemon/1/
           */
          url: string;
        };
      }[];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}